Add new packages, add routes and their pages, style homepage.
Tailwind and its prettier plugin is working. I am not sure if these are all the routes I want to have at the moment, maybe a hobby/interests just for fun too. Lastly, the basic home page is made, just need to add a few more things and style it. Probably needs to be centered vertically too.